+++ src/parser/enrichers/spell/AnimateObjects.mjs
@@ -10,14 +10,15 @@ export default class AnimateObjects extends DDBEnricherData {
 
   get activity() {
     if (this.is2014) return null;
+    const spellMod = `${this.spellModAttribute}`;
     return {
       noTemplate: true,
       profileKeys: [
-        { count: "@attributes.spellmod", name: "companion-animatedobjecttiny-2024" },
-        { count: "@attributes.spellmod", name: "companion-animatedobjectsmall-2024" },
-        { count: "@attributes.spellmod", name: "companion-animatedobjectmedium-2024" },
-        { count: "floor(@attributes.spellmod / 2)", name: "companion-animatedobjectlarge-2024" },
-        { count: "floor(@attributes.spellmod / 3)", name: "companion-animatedobjecthuge-2024" },
+        { count: `${spellMod}`, name: "companion-animatedobjecttiny-2024" },
+        { count: `${spellMod}`, name: "companion-animatedobjectsmall-2024" },
+        { count: `${spellMod}`, name: "companion-animatedobjectmedium-2024" },
+        { count: `floor(${spellMod} / 2)`, name: "companion-animatedobjectlarge-2024" },
+        { count: `floor(${spellMod} / 3)`, name: "companion-animatedobjecthuge-2024" },
       ],
       summons: {
         "match": {
@@ -44,7 +45,7 @@ export default class AnimateObjects extends DDBEnricherData {
           target: {
             affects: {
               "type": "object",
-              "count": "@attributes.spellmod",
+              "count": this.spellModAttribute,
             },
           },
         },