+++ src/parser/enrichers/spell/ProgrammedIllusion.mjs
@@ -1,0 +1,31 @@
+/* eslint-disable class-methods-use-this */
+import DDBEnricherData from "../data/DDBEnricherData.mjs";
+
+export default class ProgrammedIllusion extends DDBEnricherData {
+
+  get type() {
+    return "summon";
+  }
+
+  get activity() {
+    return {
+      noTemplate: true,
+      profileKeys: [
+        { count: 1, name: "IllusionObject" },
+        { count: 1, name: "IllusionSound" },
+        { count: 1, name: "IllusionCreature" },
+      ],
+      data: {
+        creatureSizes: [
+          "tiny",
+          "sm",
+          "med",
+          "lg",
+          "huge",
+          "grg",
+        ],
+      },
+    };
+  }
+
+}