+++ src/parser/enrichers/spell/ProjectImage.mjs
@@ -1,0 +1,24 @@
+/* eslint-disable class-methods-use-this */
+import DDBEnricherData from "../data/DDBEnricherData.mjs";
+
+export default class ProjectImage extends DDBEnricherData {
+
+  get type() {
+    return "summon";
+  }
+
+  get activity() {
+    return {
+      noTemplate: true,
+      profileKeys: [
+        { count: 1, name: "IllusionCreature" },
+      ],
+      data: {
+        creatureSizes: [
+          "med",
+        ],
+      },
+    };
+  }
+
+}